1. Рефакторинг кода 2. Полная переработка кнопки записи голосового сообщеиня

+++ app/src/main/java/com/example/messenger/utilsFilies/AppVoiceRecorder.kt
@@ -8,7 +8,6 @@ import java.io.File
 import java.io.IOException
 
 class AppVoiceRecorder {
-
     private val mMediaRecorder = MediaRecorder(mainActivityContext)
     private lateinit var mAudioFile: File
     private lateinit var mMessageKey: String
@@ -19,18 +18,23 @@ class AppVoiceRecorder {
             createAudioFile()
             prepareMediaRecorder()
         } catch (e: IOException) {
-            makeToast(e.message.toString(), mainActivityContext)
+            makeToast(e.message.toString() + "ошибка в startRecording", mainActivityContext)
         }
 
     }
 
     fun stopRecording(onSuccess: (file: File, messageKey: String) -> Unit) {
         try {
-            mMediaRecorder.stop()
-            onSuccess(mAudioFile, mMessageKey)
+            if (mAudioFile.isFile && mAudioFile.exists() && mAudioFile.length() > 0 && mMessageKey.isNotEmpty()) {
+                mMediaRecorder.stop()
+                onSuccess(mAudioFile, mMessageKey)
+            }
         } catch (e: IOException) {
-            makeToast(e.message.toString(), mainActivityContext)
-            mAudioFile.delete() //Удаляем ненужный файл
+            makeToast(e.message.toString() + "ошибка в stopRec", mainActivityContext)
+            if (mAudioFile.exists() && mAudioFile.length() > 0) {
+                mAudioFile.delete() //Удаляем ненужный файл
+            }
+
         }
     }
 
@@ -38,7 +42,7 @@ class AppVoiceRecorder {
         try {
             mMediaRecorder.release() //Удаляем экземпляр в телефоне
         } catch (e: IOException) {
-            makeToast(e.message.toString(), mainActivityContext)
+            makeToast(e.message.toString() + "ошибка в releaseRecordedVoice", mainActivityContext)
         }
     }
 
@@ -53,11 +57,11 @@ class AppVoiceRecorder {
     private fun prepareMediaRecorder() {
         mMediaRecorder.apply {
             reset() //Сбрасываем рекордер, чтобы не было конфликтов с тем, с чем он работал раньше
-            setAudioSource(MediaRecorder.AudioSource.DEFAULT) //
+            setAudioSource(MediaRecorder.AudioSource.MIC) //
             setOutputFormat(MediaRecorder.OutputFormat.DEFAULT) //выходной формат файла
             setAudioEncoder(MediaRecorder.AudioEncoder.DEFAULT) //формат кодирования файла
             setOutputFile(mAudioFile.absolutePath) //Указывыаем путь, куда сохранять аудиозапись
-            prepare() //Начинаем запись аудио
+            prepare() //Подготовливаемся к записи аудио
             start() //Запускаем запись аудио
         }
